requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.11.26=haa95532_0
certifi=2024.8.30=py312haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
contourpy=1.3.1=pypi_0
cpuonly=2.0=0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
expat=2.6.3=h5da7b33_0
filelock=3.13.1=py312haa95532_0
fonttools=4.55.0=pypi_0
freetype=2.12.1=ha860e81_0
giflib=5.2.2=h7edc060_0
idna=3.7=py312haa95532_0
imageio=2.36.1=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.4=py312haa95532_1
jpeg=9e=h827c3e9_3
kiwisolver=1.4.7=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib=3.9.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.11=py312h827c3e9_0
mkl_random=1.2.8=py312h0158946_0
moviepy=2.1.1=pypi_0
mpmath=1.3.0=py312haa95532_0
networkx=3.2.1=py312haa95532_0
numpy=2.1.3=pypi_0
numpy-base=2.0.1=py312h4dde369_1
openjpeg=2.5.2=hae555c5_0
openssl=3.0.15=h827c3e9_0
packaging=24.2=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py312haa95532_0
proglog=0.1.10=pypi_0
pyparsing=3.2.0=pypi_0
pysocks=1.7.1=py312haa95532_0
python=3.12.0=h1d929f7_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytorch=2.5.1=py3.12_cpu_0
pytorch-mutex=1.0=cpu
pyyaml=6.0.2=py312h827c3e9_0
requests=2.32.3=py312haa95532_1
setuptools=75.1.0=py312haa95532_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.2=py312haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.14=h0416ee5_0
torchaudio=2.5.1=py312_cpu
torchvision=0.20.1=py312_cpu
tqdm=4.67.1=pypi_0
typing_extensions=4.11.0=py312haa95532_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py312haa95532_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wheel=0.44.0=py312haa95532_0
win_inet_pton=1.1.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
